L2-399: e2e in testnet
